requirements.txt
langchain
langchain-community
langchain-openai
langchain-chroma
langchain-core
chromadb
tiktoken
pydantic
python-dotenv
gradio
pypdf